% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.JCR - 副本.R, R/circos.JCR.R
\name{circos.JCR}
\alias{circos.JCR}
\alias{circos.JCR}
\title{Circular visualization of Journal Citing/Cited Relationships}
\usage{
circos.JCR(x.cite, JCR.name, x.allcite, col = NULL, transparency = 0.8,
  order.by = 0, text.size = 0.7, self.include = FALSE,
  start.degree = 90, draw.line = FALSE, ...)

circos.JCR(x.cite, JCR.name, x.allcite, col = NULL, transparency = 0.8,
  order.by = 0, text.size = 0.7, self.include = FALSE,
  start.degree = 90, draw.line = FALSE, ...)
}
\arguments{
\item{JCR.name}{Journal names (N.B., can NOT rename, i.e. No name is repeated each other).}

\item{col}{The color for sector.index and circos.link.}

\item{transparency}{Transparency of color.}

\item{order.by}{The direction of the circos, 0 = orde by the number of 'x', 1 = order by alphabet for 'JCR.name'.}

\item{text.size}{The font size of text. Value as cex (e.g., 0.8).}

\item{self.include}{Whether the link include the parent journal. Default is FALSE.}

\item{start.degree}{For the start degree of the 1st sector.index. see \code{\link[circlize]{circos.par}}.}

\item{...}{additional parameters to \code{\link[circlize]{circos.link}},such as directional, rou, rou1, tou2, lty, arr.lty.}

\item{x}{Citing or cited data of JCR in sector.index1. The 1st value in dataset 'x' is the sector.index1 as default.}

\item{track.height}{The height of one track.index.}

\item{x}{Citing or cited data of JCR in sector.index1. The 1st value in dataset 'x' is the sector.index1 as default.}

\item{JCR.name}{Journal names (N.B., can NOT rename, i.e. No name is repeated each other).}

\item{col}{The color for sector.index and circos.link.}

\item{transparency}{Transparency of color.}

\item{order.by}{The direction of the circos, 0 = orde by the number of 'x', 1 = order by alphabet for 'JCR.name'.}

\item{self.include}{Whether the link include the parent journal. Default is FALSE.}

\item{start.degree}{For the start degree of the 1st sector.index. see \code{\link[circlize]{circos.par}}.}

\item{text.size}{The font size of text. Value as cex (e.g., 0.8).}

\item{track.height}{The height of one track.index.}

\item{...}{additional parameters to \code{\link[circlize]{circos.link}},such as directional, rou, rou1, tou2, lty, arr.lty.}
}
\description{
circos.JCR is a extension of 'circlize' package for drawing the graph of Journal Citing/Cited Relationships.
Graph example as http://ipscience-help.thomsonreuters.com/incitesLiveJCR/JCRGroup/jcrJournalProfile/jcrJournalProfileEgoNetwork.html

circos.JCR is a extension of 'circlize' package for drawing the graph of Journal Citing/Cited Relationships.
Graph example as http://ipscience-help.thomsonreuters.com/incitesLiveJCR/JCRGroup/jcrJournalProfile/jcrJournalProfileEgoNetwork.html
}
\examples{

library(circos.JCR)
# an example of citing data of journal "Nature", the top 20 cited journals data included.
NatureCiting = c(4665, 2719, 1757, 1279, 713, 630, 556, 445, 441, 413,
                 402,  387,  367,  359,  351, 337, 323, 316, 314, 312, 292)
# Top 20 cited journals by the parent journal "Nature"
JCR.name =c("Nature", "SCIENCE", "P NATL ACAD SCI USA", "CELL", "PHYS REV LETT",
            "ASTROPHYS J", "J BIOL CHEM", "NAT COMMUN", "NUCLEIC ACIDS RES", "PLOS ONE",
             "NEURON", "J AM CHEM SOC", "J NEUROSCI", "BIOINFORMATICS", "MON NOT R ASTRON.",
             "ACTA CRYSTALLOGR D", "NAT METHODS", "MOL CELL", "NAT GENET", "ASTRON ASTROPHYS", "NAT NEUROSCI")
# "Citing data (not exact data)" by the top 20 journals.
Allciting = c(7156,  6564,  12214, 2818, 968,
              19399, 7772, 10489, 6570, 27664,
              2251,  3487, 6329,  1475, 28571,
              501, 600, 1198, 668, 10747, 2460)

circos.JCR(x.cite = NatureCiting, JCR.name = JCR.name, x.allcite = Allciting, start.degree = 90, order.by = 0, text.size = 0.7)


library(circos.JCR)
NatureCiting = c(4665, 2719, 1757, 1279, 713, 630, 556, 445, 441, 413,
                 402,  387,  367,  359,  351)
JCR.name =c("Nature", "SCIENCE", "P NATL ACAD SCI USA", "CELL", "PHYS REV LETT",
            "ASTROPHYS J", "J BIOL CHEM", "NAT COMMUN", "NUCLEIC ACIDS RES", "PLOS ONE",
             "NEURON", "J AM CHEM SOC", "J NEUROSCI", "BIOINFORMATICS", "MON NOT R ASTRON.")
Allciting = c(7156,  6564,  12214, 2818, 968,
              19399, 7772, 10489, 6570, 27664,
              2251,  3487, 6329,  1475, 28571)

circos.JCR(x.cite = NatureCiting, JCR.name = JCR.name, x.allcite = Allciting, start.degree = 90, order.by = 0, text.size = 0.7)

}
\seealso{
\code{\link[circlize]{circos.link}}

\code{\link[circlize]{circos.link}}
}
\author{
Weiping Mei <meiweipingg@163.com>

Weiping Mei <meiweipingg@163.com>
}
