% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.JCR.R
\name{circos.JCR}
\alias{circos.JCR}
\title{Circular visualization of Journal Citing/Cited Relationships}
\usage{
circos.JCR(x, JCR.name, start.degree = 90, col = 1:length(x),
  order.by = 0, text.facing = "outside", text.size = 0.5, ...)
}
\arguments{
\item{x}{Citing or cited data of JCR in sector.index1. The 1st value in dataset 'x' is the sector.index1 as default.}

\item{JCR.name}{Journal names (N.B., can NOT rename, i.e. No name is repeated each other).}

\item{start.degree}{For the start degree of the 1st sector.index. see \code{\link[circlize]{circos.par}}.}

\item{col}{The color for sector.index and circos.link.}

\item{order.by}{The direction of the circos, 0 = orde by the number of 'x', 1 = order by alphabet for 'JCR.name'.}

\item{text.facing}{The value for 'facing' in \code{\link[circlize]{circos.text}}.}

\item{text.size}{The font size of text. Value as cex (e.g., 0.8).}

\item{...}{additional parameters to \code{\link[circlize]{circos.link}},such as directional, lty, arr.lty.}
}
\description{
circos.JCR is a extension of 'circlize' package for drawing the graph of Journal Citing/Cited Relationships. 
Graph example as http://ipscience-help.thomsonreuters.com/incitesLiveJCR/JCRGroup/jcrJournalProfile/jcrJournalProfileEgoNetwork.html
}
\examples{

library(circlize)
library(circos.JCR)
x = c(15, 11, 18, 16, 14, 13, 12, 7)  # an example of citing data of journal "Nature"
Name =c("Nature", "PLOS ONE", "SCIENCE", "CELL", "ECOLOGY", "OTHER1", "OTHER2", "OTHER3")

circos.JCR(x = x, JCR.name = Name, start.degree = 90, order.by = 0, text.size = 0.7)

}
\seealso{
\code{\link[circlize]{circos.link}}
}
\author{
Weiping Mei <meiweipingg@163.com>
}
